requirements.txt
altair==5.1.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
blis==0.7.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
confection==0.1.0
cryptography==41.0.1
ctparse==0.3.6
cymem==2.0.7
dnspython==2.3.0
filelock==3.12.2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-PyMongo==2.3.0
fsspec==2023.6.0
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.37
hrflow==3.0.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonmerge==1.9.0
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
langcodes==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
numpy==1.25.1
packaging==23.1
pandas==2.1.1
pathy==0.10.2
Pillow==10.0.1
preshed==3.0.8
protobuf==4.24.4
pyarrow==13.0.0
pycparser==2.21
pydantic==1.10.11
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.7.0
pymongo==4.4.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rich==13.6.0
rpds-py==0.8.10
safetensors==0.3.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
streamlit==1.27.2
sympy==1.12
tenacity==8.2.3
thinc==8.1.10
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.3.3
tqdm==4.65.0
transformers==4.30.2
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.3
validators==0.22.0
wasabi==1.1.2
watchdog==3.0.0
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.16.0
